// berger-one/src/services/constants/interfaces/visit.interface.ts
export interface I_Visit_Details {
    dcio_id: any;
    dcio_depot_code: any;
    leadId: any;
    dcio_check_in_time: any;
    dcio_check_out_time: any;
    dcio_activity_others: any;
    dcio_activity_type: any;
    dcio_tsi_visit_made_yn: any;
    dcio_tsi_visit_feedback: any;
    dcio_tsi_visit_remarks: any;
    dcio_propect_visit_made_yn: any;
    dcio_prospect_visit_feedback: any;
    dcio_prospect_visit_remarks: any;
    dcio_geo_tag_not_done_reason: any;
    dcio_visit_next_action: any;
    dcio_visit_by: any;
    dcio_conflict_dealer_code: any;
    dcio_conflict_dealer_distance: any;
    dcio_conflict_dealer_business_value: any;
    dcio_cancelation_request_yn: any;
    dcio_cancelation_request_remarks: any;
    dcio_next_followup_date: any;
    nonServiceablePin: any;
}

export const Visit_Details_DTO: I_Visit_Details = {
    dcio_id: null,
    dcio_depot_code: null,
    leadId: null,
    dcio_check_in_time: null,
    dcio_check_out_time: null,
    dcio_activity_others: null,
    dcio_activity_type: null,
    dcio_tsi_visit_made_yn: 'Y',
    dcio_tsi_visit_feedback: null,
    dcio_tsi_visit_remarks: null,
    dcio_propect_visit_made_yn: 'Y',
    dcio_prospect_visit_feedback: null,
    dcio_prospect_visit_remarks: null,
    dcio_geo_tag_not_done_reason: null,
    dcio_visit_next_action: null,
    dcio_visit_by: null,
    dcio_conflict_dealer_code: null,
    dcio_conflict_dealer_distance: null,
    dcio_conflict_dealer_business_value: null,
    dcio_cancelation_request_yn: 'N',
    dcio_cancelation_request_remarks: null,
    dcio_next_followup_date: null,
    nonServiceablePin: null,
};

export interface I_Visit_History {
    dcio_id: any;
    dcio_depot_code: any;
    dcio_dealer_code: any;
    dcio_booking_id: any;
    dcio_check_in_long: any;
    dcio_check_in_lat: any;
    dcio_check_in_time: any;
    dcio_check_in_accuracy: any;
    dcio_check_out_long: any;
    dcio_check_out_lat: any;
    dcio_check_out_time: any;
    dcio_check_out_accuracy: any;
    dcio_spent_min: any;
    dcio_activity_others: any;
    created_user: any;
    created_date: any;
    modified_user: any;
    modified_date: any;
    deleted_user: any;
    deleted_date: any;
    active: any;
    dcio_activity_type: any;
    sync_ref_id: any;
    dcio_tsi_visit_made_yn: any;
    dcio_tsi_visit_made_on: any;
    dcio_tsi_visit_made_by: any;
    dcio_tsi_visit_feedback: any;
    dcio_tsi_visit_remarks: any;
    dcio_propect_visit_made_yn: any;
    dcio_propect_visit_made_on: any;
    dcio_propect_visit_made_by: any;
    dcio_prospect_visit_feedback: any;
    dcio_prospect_visit_remarks: any;
    dcio_visit_discrepancy_yn: any;
    dcio_visit_discrepancy_on: any;
    dcio_visit_discrepancy_by: any;
    dcio_visit_discrepancy_raised_yn: any;
    dcio_visit_discrepancy_raised_on: any;
    dcio_visit_discrepancy_raised_by: any;
    dcio_geo_tag_not_done_reason: any;
    dcio_visit_next_action: any;
    dcio_further_appointment_taken_yn: any;
    dcio_tsi_visit_feedback_value: any;
    dcio_prospect_visit_feedback_value: any;
    dcio_visit_next_action_value: any;
}